@@ -10,7 +10,9 @@
         "lint": "eslint --fix -c ./.eslintrc.js \"src/**/*.ts*\"",
         "prettify": "prettier --write \"src/**/*.ts*\"",
         "storybook": "start-storybook -p 6006",
-        "build-storybook": "build-storybook"
+        "build-storybook": "build-storybook",
+        "dev-vite": "vite",
+        "build-vite" : "vite build"
     },
     "repository": {
         "type": "git",
@@ -42,9 +44,9 @@
         "@types/chrome": "^0.0.124",
         "@types/jest": "^26.0.14",
         "@types/node": "^14.11.8",
-        "@types/react": "^17.0.2",
-        "@types/react-dom": "^17.0.2",
-        "@types/react-test-renderer": "^17.0.1",
+        "@types/react": "^18.0.0",
+        "@types/react-dom": "^18.0.0",
+        "@types/react-test-renderer": "^18.0.0",
         "@types/webextension-polyfill": "^0.9.0",
         "@typescript-eslint/eslint-plugin": "^4.4.1",
         "@typescript-eslint/parser": "^4.4.1",
@@ -63,18 +65,21 @@
         "postcss": "^8.4.5",
         "postcss-loader": "^6.2.1",
         "prettier": "^2.1.2",
-        "react-test-renderer": "^17.0.2",
+        "react-test-renderer": "^18.0.0",
         "tailwindcss": "^3.0.8",
         "ts-jest": "^26.4.1",
         "ts-loader": "^8.0.5",
-        "typescript": "^4.0.3",
+        "typescript": "^4.6.3",
         "webpack": "^5.65.0",
         "webpack-cli": "^4.9.1",
         "webpack-merge": "^5.8.0"
     },
     "dependencies": {
-        "react": "^17.0.2",
-        "react-dom": "^17.0.2",
+        "@vitejs/plugin-react": "^1.3.0",
+        "react": "^18.0.0",
+        "react-dom": "^18.0.0",
+        "vite": "^2.9.9",
+        "vite-plugin-chrome-extension": "^0.0.7",
         "webextension-polyfill": "^0.9.0"
     }
 }
